% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flusher.R
\name{flush_editorcontent}
\alias{flush_editorcontent}
\title{Flush the content of the editor window (even if not saved)}
\usage{
flush_editorcontent(out_suffix = ".flushed.R")
}
\arguments{
\item{out_suffix}{Character string, defaults to .flushed.R}
}
\value{
Creates a file
}
\description{
Flush the content of the editor window (even if not saved), saving it into a file
whose name is the same as the filename in the editor, with a specified suffix
}
\examples{
# flush_editorcontent()     # e.g. to be inserted in the execution of a scripts
}
