% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_pm.R
\name{get_corr_pairedTtest}
\alias{get_corr_pairedTtest}
\title{Get correlation of paired groups from summary statistics}
\usage{
get_corr_pairedTtest(meanDiff, sd1, sd2, pValue)
}
\arguments{
\item{meanDiff}{Difference of means in two paired groups}

\item{sd1}{Standard deviation in group 1}

\item{sd2}{Standard deviation in group 2}

\item{pValue}{p value as reported in paper}
}
\value{
the correlation of the two groups
}
\description{
Sometimes, papers report mean and std. dev of paired groups along with a p-value,
but not the correlation (Often done in a typical table 1). The latter is needed
for sample size calculations and can be recovered with this function.
}
