% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_pm.R
\name{print_with_confint}
\alias{print_with_confint}
\title{print_with_confint}
\usage{
print_with_confint(coefs, df = Inf, exp = FALSE, conf.level = 0.95)
}
\arguments{
\item{coefs}{two-column matrix. First column should be the estimate, second column should be the the standard error as returned from summary method}

\item{df}{numeric, denominator degrees of freedom if t-d}

\item{exp}{logical, defaults to FALSE. Should the estimates be transformed with exp(.)?}

\item{conf.level}{numeric, confidence level. Defaults to 0.95 .}
}
\value{
a matrix
}
\description{
this is a small auxiliary function that reformats the coefficient estimates of regression models.
}
\examples{

data(datasets::mtcars)
mod   <- lm(mpg ~ disp+hp+wt, data=mtcars)
coefs <- summary(mod)$coefficients[,1:2] 

print_with_confint(coefs)

}
