% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCking.R
\name{rocplot}
\alias{rocplot}
\title{Customized ROC plot}
\usage{
rocplot(df, predictions_name, outcomes_name, decorate_youden = TRUE,
  decorate_performance = TRUE, ...)
}
\arguments{
\item{df}{The data frame containing the observations}

\item{predictions_name}{Character, specifying the column name containing the predictor}

\item{outcomes_name}{Character, specifying the column name specifying the labels}

\item{decorate_youden}{Logical, whether to add some info about and around the Youden Index}

\item{decorate_performance}{Logical, whether to add some info about the performance
of the classifier}

\item{...}{To be further passed to the ROCR::plot command, ensures some degree of
customizability (color, add=TRUE, ...)}
}
\value{
A plot object, and invisibly the list of objects which can be conveniently
further used
}
\description{
Customized ROC plot
}
\examples{
# Using the basic example from the ROCR package
library(ROCR)
data(ROCR.simple)
pred <- prediction( ROCR.simple$predictions, ROCR.simple$labels )
perf <- performance( pred, "tpr", "fpr" )
plot( perf, main ="made with default plot function from ROCR" )

rocplot(ROCR.simple,"predictions","labels", main = "made with rocplot")
}
