% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_imbei_beamer.R
\name{use_imbei_beamer}
\alias{use_imbei_beamer}
\title{use_imbei_beamer}
\usage{
use_imbei_beamer(keep_tex = TRUE, copy_to = getwd())
}
\arguments{
\item{keep_tex}{Logical, defines wheter the tex file is kept. Defaults to TRUE}

\item{copy_to}{File path for beamer template. Defaults to getwd()}
}
\value{
Creates a tex and prints out the suggested yaml content
}
\description{
Facilitates the use of the IMBEI beamer template.
This is an alternative to creating a new markdown from template directly.
Latter is preferred.
}
\examples{
use_imbei_beamer()
}
