% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMBEImisc-pkg.R
\docType{package}
\name{IMBEImisc-pkg}
\alias{IMBEImisc-pkg}
\alias{IMBEImisc-pkg-package}
\title{IMBEImisc}
\description{
\code{IMBEImisc} is an R package that provides
}
\author{
Federico Marini \email{marinif@uni-mainz.de}
}
