% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_pm.R
\name{plot_pkg}
\alias{plot_pkg}
\title{plot_pkg}
\usage{
plot_pkg(pkg_names, start_date = "2012-01-01", ...)
}
\arguments{
\item{pkg_names}{character string of package names to compare}

\item{start_date}{character in ISO-format, default is "2012-01-01"}

\item{...}{}
}
\value{
a ggplot-object
}
\description{
A function to check and compare the "poplarity" of R-packages.
It uses the cranlogs-package to get download counts, which are then plotted.
}
\examples{
plot_pkg(c("gee","geepack"))


}
