% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_imbei_beamer.R
\name{use_imbei_beamer}
\alias{use_imbei_beamer}
\title{use_imbei_beamer}
\usage{
use_imbei_beamer(keep_tex = TRUE, copy_to = getwd())
}
\arguments{
\item{keep_tex}{Logical, fffd}

\item{copy_to}{Path whehejhhe}
}
\value{
Creates a tex and prints out the suggested yaml content
}
\description{
compact descr
}
\details{
some more details
}
\examples{
use_imbei_beamer()
}
